Add files via upload
The project “Online Election System” / “Online Voting Software” aims at making the voting process easy in any type of elections. Presently voting is performed using ballot paper and the counting is done manually, hence it consumes a lot of time. There can be possibility of invalid votes. All these make election a tedious task. In recent times in India, due to elections the second wave of COVID transmission also made huge loss of human lives. In our proposed system voting and counting is done with the help of computer in Online It saves time, avoid error in counting and there will be no invalid votes. It makes the election process easy. It also avoids the process of physical touching or visiting any places and so in the time of pandemic too it will be more helpful to conduct elections